+++ src/commands/runAll.js
@@ -2,5 +2,8 @@ const { runInTerminal } = require("../terminal");
 const helpers = require("../helpers");
 
 module.exports = {
-    runAll: () => runInTerminal(helpers.getSetting("runAllCommand"))
+    runAll: () => {
+        const command = helpers.getSetting("runAllCommand").replace("{base}", helpers.getSetting("baseCommand"));
+        runInTerminal(command);
+    }
 };