+++ src/commands/runFile.js
@@ -4,6 +4,7 @@ const helpers = require("../helpers");
 module.exports = {
     runFile: () => {
         const command = helpers.getSetting("runFileCommand")
+            .replace("{base}", helpers.getSetting("baseCommand"))
             .replace("{fileName}", helpers.getActiveFile());
 
         runInTerminal(command);